Build the filters object in ToolsLayout
